pyproject.toml
[project]
name = "cman"
version = "0.1.0"
description = "manage mochi cards"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "Pillow",
    "debugpy",
    "flask",
    "ipdb",
    "ipython",
    "pandoc",
    "pydantic",
    "pyserde[toml]",
    "requests",
    "tqdm",
    "typer",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cman"]

[project.scripts]
cman = "cman.cli:app"

[tool.pyright]
include = ["src"]
typeCheckingMode = "standard"
reportUnreachable = false
